Refactor utility functions exports
